tests/tests/docker/test_production_deployment.py
"""
Phase 2 Docker Container Tests - Additional Implementation

This module provides comprehensive Docker container testing for the Integral Philosophy Publishing System,
including container orchestration, production deployment scenarios, and service dependency validation.
"""

import pytest
import time
import json
import requests
import tempfile
import subprocess
import yaml
import os
import socket
from pathlib import Path
from typing import Dict, List, Any, Optional, Tuple
from unittest.mock import Mock, patch, MagicMock

from ..utils.base_test_classes import BaseTestCase, IntegrationTestCase


class TestContainerOrchestration(IntegrationTestCase):
    """Test Docker Compose orchestration without requiring Docker daemon"""

    @pytest.fixture(autouse=True)
    def setup_orchestration_environment(self):
        """Setup orchestration testing environment"""
        self.compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"
        self.compose_cmd = ["docker-compose", "-f", str(self.compose_file)]

    def test_compose_file_validation(self):
        """Validate Docker Compose file structure and content"""
        if not self.compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(self.compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        # Validate version
        assert "version" in compose_config, "Missing version in docker-compose.yml"

        # Validate services
        services = compose_config.get("services", {})
        assert len(services) > 0, "No services defined"

        # Common services to check for
        expected_services = ["web", "db", "redis"]
        found_services = []

        for service in expected_services:
            if service in services:
                found_services.append(service)

        if len(found_services) == 0:
            pytest.skip("No expected services found in docker-compose.yml")

    def test_service_configuration_validation(self):
        """Validate individual service configurations"""
        if not self.compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(self.compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        # Validate web service if present
        if "web" in services:
            web_service = services["web"]

            # Check for port exposure
            if "ports" in web_service:
                ports = web_service["ports"]
                assert len(ports) > 0, "Web service should expose ports"

            # Check for environment variables
            if "environment" in web_service:
                env = web_service["environment"]
                if isinstance(env, dict):
                    # Check for required environment variables
                    required_env = ["DATABASE_URL", "REDIS_URL"]
                    for env_var in required_env:
                        assert env_var in env, f"Web service missing {env_var}"

        # Validate database service if present
        if "db" in services:
            db_service = services["db"]

            # Check for volume mounting for persistence
            if "volumes" in db_service:
                volumes = db_service["volumes"]
                assert len(volumes) > 0, "Database should have persistent volumes"

            # Check for environment variables
            if "environment" in db_service:
                env = db_service["environment"]
                if isinstance(env, dict):
                    # Check for database credentials
                    db_vars = ["POSTGRES_USER", "POSTGRES_PASSWORD", "POSTGRES_DB"]
                    for var in db_vars:
                        assert var in env, f"Database missing {var}"


class TestContainerLifecycle(BaseTestCase):
    """Test container lifecycle management"""

    def test_startup_order_validation(self):
        """Test service startup order dependencies"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        # Validate dependency relationships
        for service_name, service_config in services.items():
            depends_on = service_config.get("depends_on", [])

            for dependency in depends_on:
                assert dependency in services, (
                    f"Service {service_name} depends on non-existent service: {dependency}"
                )

    def test_health_check_implementation(self):
        """Test health check implementations"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        for service_name, service_config in services.items():
            health_check = service_config.get("healthcheck")

            if health_check:
                assert "test" in health_check, (
                    f"Service {service_name} health check missing test command"
                )


class TestContainerSecurity(BaseTestCase):
    """Test container security configurations"""

    def test_dockerfile_security(self):
        """Test Dockerfile for security best practices"""
        dockerfile_path = Path(__file__).parent.parent.parent / "Dockerfile"

        if not dockerfile_path.exists():
            pytest.skip("Dockerfile not found")

        with open(dockerfile_path, "r") as f:
            dockerfile_content = f.read()

        # Security checks
        security_issues = []

        # Check for non-root user
        if "USER" not in dockerfile_content:
            security_issues.append(
                "Dockerfile should specify a non-root user with USER instruction"
            )

        # Report security issues
        if security_issues:
            print("Security recommendations:\n" + "\n".join(security_issues))


class TestContainerNetworking(BaseTestCase):
    """Test container networking configurations"""

    def test_network_isolation(self):
        """Test network isolation configurations"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        networks = compose_config.get("networks", {})
        services = compose_config.get("services", {})

        # Validate network references
        for service_name, service_config in services.items():
            service_networks = service_config.get("networks", [])

            if isinstance(service_networks, list):
                for network in service_networks:
                    if network not in networks:
                        pytest.fail(
                            f"Service {service_name} references undefined network: {network}"
                        )

    def test_port_exposure_security(self):
        """Test secure port exposure configurations"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        # Check for database port exposure
        for service_name, service_config in services.items():
            if service_name in ["db", "database", "postgres", "mysql"]:
                ports = service_config.get("ports", [])
                if ports:
                    pytest.fail(
                        f"Database service {service_name} should not expose ports externally"
                    )


class TestContainerPerformance(BaseTestCase):
    """Test container performance configurations"""

    def test_resource_limits(self):
        """Test resource limit configurations"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        # Check for resource limits on critical services
        for service_name in ["web", "api"]:
            if service_name in services:
                service_config = services[service_name]
                deploy_config = service_config.get("deploy", {}).get("resources", {})

                if not deploy_config.get("limits", {}).get("memory"):
                    print(f"Service {service_name} missing memory limit")


class TestContainerMonitoring(BaseTestCase):
    """Test container monitoring configurations"""

    def test_logging_configuration(self):
        """Test logging configuration for containers"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        # Check for logging configuration
        for service_name, service_config in services.items():
            logging_config = service_config.get("logging")

            if not logging_config:
                print(f"Service {service_name} uses default logging configuration")

    def test_metrics_collection(self):
        """Test metrics collection setup"""
        compose_file = Path(__file__).parent.parent.parent / "docker-compose.yml"

        if not compose_file.exists():
            pytest.skip("docker-compose.yml not found")

        try:
            with open(compose_file, "r") as f:
                compose_config = yaml.safe_load(f)
        except yaml.YAMLError as e:
            pytest.fail(f"Invalid docker-compose.yml: {e}")

        services = compose_config.get("services", {})

        # Look for monitoring services
        monitoring_services = ["prometheus", "grafana", "influxdb"]
        found_monitoring = []

        for service_name in services:
            if any(monitor in service_name.lower() for monitor in monitoring_services):
                found_monitoring.append(service_name)

        if found_monitoring:
            print(f"Found monitoring services: {', '.join(found_monitoring)}")
        else:
            print("No monitoring services found in docker-compose.yml")
