tests/requirements.txt
# Core testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-html>=4.1.0
pytest-mock>=3.11.1
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1

# Selenium for web testing
selenium>=4.15.0
webdriver-manager>=4.0.1

# HTTP testing
requests-mock>=1.11.0
httpx>=0.24.1

# Database testing (if needed)
pytest-postgresql>=5.0.0
pytest-redis>=3.0.2

# Performance testing
pytest-benchmark>=4.0.0
locust>=2.17.0

# Security testing
bandit>=1.7.5
safety>=2.3.5

# Coverage and reporting
coverage>=7.3.2
codecov>=2.1.13

# Async testing support
aiohttp>=3.8.6
asynctest>=0.13.0

# Mocking and fixtures
responses>=0.23.3
factory-boy>=3.3.0
faker>=20.0.0

# Docker testing (for integration)
pytest-docker>=2.0.1
docker>=6.1.3

# Parallel execution
pytest-parallel>=0.1.1

# Test data generation
hypothesis>=6.87.1
parameterized>=0.9.0

# Report generation
allure-pytest>=2.13.2
pytest-json-report>=1.5.0

# Environment management
python-dotenv>=1.0.0
pytest-env>=1.1.1

# Memory profiling
memory-profiler>=0.61.0
pytest-memray>=1.5.0

# Network testing for scraping
mitmproxy>=10.1.5
freezegun>=1.2.2

# File handling and comparison
filecmp>=0.2
deepdiff>=6.7.1

# Logging and debugging
pytest-logger>=0.2.0
pytest-sugar>=0.9.7

# Type checking support
mypy>=1.7.0
pytest-mypy-plugins>=3.0.0